+++ BasicMemoryInterface.h
@@ -13,10 +13,10 @@ public:
     ~BasicMemoryInterface();
 
     bool UpdateProcessId(const wchar_t* processName);
-    intptr_t GetBaseAddress();
-    bool ReadRaw(intptr_t address, void* pBuffer, unsigned long size);
-    bool WriteRaw(intptr_t address, void* pBuffer, unsigned long size);
-    intptr_t GetModuleBase();
+    uint64_t GetBaseAddress();
+    bool ReadRaw(uint64_t address, void* pBuffer, unsigned long size);
+    bool WriteRaw(uint64_t address, void* pBuffer, unsigned long size);
+    uint64_t GetModuleBase();
     bool SetTargetModule(wchar_t* moduleName);
 private:
     void* handle;
