+++ DriverInterop.cpp
@@ -14,24 +14,24 @@
 
 struct IOReadRequest {
     unsigned long ProcessId;
-    intptr_t Address;
+    uint64_t Address;
     SIZE_T Size;
     unsigned char* Value;
 };
 
 struct IOWriteRequest {
     unsigned long ProcessId;
-    intptr_t Address;
+    uint64_t Address;
     SIZE_T Size;
     unsigned char* Value;
 };
 
 struct IOGetBaseAddressRequest {
     DWORD ProcessId;
-    intptr_t BaseAddress;
+    uint64_t BaseAddress;
 };
 
-bool DriverInterop::ReadRaw(intptr_t address, void* data, unsigned long size) {
+bool DriverInterop::ReadRaw(uint64_t address, void* data, unsigned long size) {
     if (address < MINIMUM_ADDRESS_SIZE) {
         return false;
     }
@@ -41,7 +41,7 @@ bool DriverInterop::ReadRaw(intptr_t address, void* data, unsigned long size) {
     return DeviceIoControl(this->handle, IO_READ_REQUEST, &request, sizeof(IOReadRequest), &request, sizeof(IOReadRequest), nullptr, nullptr);
 }
 
-bool DriverInterop::WriteRaw(intptr_t address, void* data, unsigned long size) {
+bool DriverInterop::WriteRaw(uint64_t address, void* data, unsigned long size) {
     if (address < MINIMUM_ADDRESS_SIZE) {
         return false;
     }
@@ -61,7 +61,7 @@ bool DriverInterop::UpdateProcessId(const wchar_t* processName) {
     return this->pid != 0;
 }
 
-intptr_t DriverInterop::GetBaseAddress() {
+uint64_t DriverInterop::GetBaseAddress() {
     IOGetBaseAddressRequest request{ this->pid, 0 };
 
     return 
@@ -70,16 +70,16 @@ intptr_t DriverInterop::GetBaseAddress() {
             &request, sizeof(IOGetBaseAddressRequest),
             &request, sizeof(IOGetBaseAddressRequest),
             nullptr, nullptr
-        ) ? (intptr_t)request.BaseAddress : 0;
+        ) ? (uint64_t)request.BaseAddress : 0;
 }
 
-intptr_t DriverInterop::GetModuleBase() {
-    intptr_t request = 0;
+uint64_t DriverInterop::GetModuleBase() {
+    uint64_t request = 0;
 
     return DeviceIoControl(this->handle,
         IO_GET_MODULE_BASE_REQUEST,
-        &request, sizeof(intptr_t),
-        &request, sizeof(intptr_t),
+        &request, sizeof(uint64_t),
+        &request, sizeof(uint64_t),
         nullptr, nullptr
     ) ? request : 0;
 }
