+++ GameWorld.cpp
@@ -3,12 +3,12 @@
 #include "Global.hpp"
 
 GameWorld GameWorld::Get() {
-	intptr_t gameWorldAddressGameObject = Global::gom.GetActiveObjects().GetObject("GameWorld");
+	uint64_t gameWorldAddressGameObject = Global::gom.GetActiveObjects().GetObject("GameWorld");
 	if (!gameWorldAddressGameObject) {
 		return GameWorld{0};
 	}
 
-	intptr_t gameWorldAddress = Memory::ReadChain<intptr_t>(
+	uint64_t gameWorldAddress = Memory::ReadChain<uint64_t>(
 		Global::pMemoryInterface,
 		gameWorldAddressGameObject,
 		{0x30, 0x18, 0x28}
@@ -20,12 +20,12 @@ GameWorld GameWorld::Get() {
 std::vector<Player> GameWorld::GetPlayers() {
 	if (!this->cachedPlayers.has_value()) {
 		std::vector<Player> players{};
-		intptr_t registeredPlayersAddress = Memory::ReadValue<intptr_t>(
+		uint64_t registeredPlayersAddress = Memory::ReadValue<uint64_t>(
 			Global::pMemoryInterface,
 			this->address + Offsets::GameWorld::RegisteredPlayers
 		);
 		if (registeredPlayersAddress) {
-			std::tuple<int, intptr_t*> playerAddresses = Memory::ReadList<intptr_t>(Global::pMemoryInterface, registeredPlayersAddress);
+			std::tuple<int, uint64_t*> playerAddresses = Memory::ReadList<uint64_t>(Global::pMemoryInterface, registeredPlayersAddress);
 			for (int i = 0; i < std::get<0>(playerAddresses); i++) {
 				players.push_back(Player{std::get<1>(playerAddresses)[i]});
 			}
@@ -39,9 +39,9 @@ std::vector<Player> GameWorld::GetPlayers() {
 
 std::vector<WorldLootItem> GameWorld::GetLoot() {
 	if (!this->cachedLoot.has_value()) {
-		std::tuple<int, intptr_t*> list = Memory::ReadList<intptr_t>(Global::pMemoryInterface, Memory::ReadValue<intptr_t>(Global::pMemoryInterface, this->address + 0x80));
+		std::tuple<int, uint64_t*> list = Memory::ReadList<uint64_t>(Global::pMemoryInterface, Memory::ReadValue<uint64_t>(Global::pMemoryInterface, this->address + 0x80));
 		int count = std::get<0>(list);
-		intptr_t* addresses = std::get<1>(list);
+		uint64_t* addresses = std::get<1>(list);
 		std::vector<WorldLootItem> loot{};
 		loot.reserve(count);
 		for (int i = 0; i < count; i++) {
