+++ ItemTemplate.hpp
@@ -6,14 +6,14 @@
 
 class ItemTemplate {
 public:
-	ItemTemplate(intptr_t address) {
+	ItemTemplate(uint64_t address) {
 		this->address = address;
 	}
 
 	std::wstring GetId();
 	std::wstring GetLocalizedName();
 
-	intptr_t address;
+	uint64_t address;
 
 private:
 	std::optional<std::wstring> cachedId;
