+++ Slot.hpp
@@ -9,14 +9,14 @@ public:
 	Slot() : Slot(0) {
 	}
 
-	Slot(intptr_t address) {
+	Slot(uint64_t address) {
 		this->address = address;
 	}
 
 	Item GetContainedItem();
 	bool IsRequired();
 
-	intptr_t address;
+	uint64_t address;
 
 private:
 	std::optional<Item> cachedContainedItem = std::nullopt;
